refactor(doctor): share preview warning rendering
